Add scripts for client-server workflow.
- ems starts emacs with a server
- eml lists running servers
- emc connects to a server

GUI can only be a standalone instance, e.g., not a client, thus emc only
starts in text mode.

The reason for this is that I was once hit by a bug, where I ran a daemon
server and had a GUI client connected to it. I left it for a while, only
to find the GUI client dead (while the X window was not deleted, leaving
a undrawn frame), and the daemon spinning 100% CPU and unresponding to
any new clients.

But still, I'd like to have server mode, in case I need to ssh to a
machine and re-attach running emacs instances. For emacs instances that
are intended to live long, I would start them with ems.